Final touches to color
- Create an edit component file
- Tie editor colors to prop.color for consistent output between default color, custom color, and theme defined color
