+++ src/blocks/block-newsletter/index.js
@@ -1,126 +1,15 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-import * as colorClassSlug from './../../components/color-class';
-
-/**
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
 const { registerBlockType } = wp.blocks;
-const { getColorClassName, RichText } = wp.editor;
-const { Fragment, Component } = wp.element;
-const { TextControl } = wp.components;
 
 /**
  * Internal dependencies
  */
-import Inspector from './components/inspector';
-import CustomButton from './../block-button/components/button';
-import NewsletterContainer from './components/newsletter';
+import Edit from './components/edit';
 import './styles/style.scss';
 import './styles/editor.scss';
-
-class Edit extends Component {
-	constructor() {
-		super( ...arguments );
-	}
-
-	render() {
-		const {
-			attributes,
-			isSelected,
-			setAttributes,
-		} = this.props;
-
-		const apiKeyDefined = atomic_blocks_newsletter_block_vars.mailingListProviders.mailchimp.api_key_defined;
-
-		const editClassName = classnames( {
-			'ab-block-button': true,
-		} );
-
-		const editStyles = {
-			backgroundColor: attributes.backgroundColor,
-		};
-
-		const getButtonTextClass       = getColorClassName( 'color', attributes.buttonTextColor );
-		const getButtonBackgroundClass = getColorClassName( 'background-color', attributes.buttonBackgroundColor );
-
-		return [
-			<Inspector { ...{ setAttributes, ...this.props } }/>,
-			<NewsletterContainer { ...this.props }>
-				{ ! apiKeyDefined && (
-					<Fragment>
-						<div className="ab-newsletter-notice">
-							{ __( 'You must define your newsletter provider API keys to use this block.', 'atomic-blocks' ) }
-							<p><a href={ atomic_blocks_newsletter_block_vars.plugin_settings_page_url }>{ __( 'Configure your settings', 'atomic-blocks' ) }</a></p>
-						</div>
-					</Fragment>
-				) }
-				{ apiKeyDefined && (
-					<Fragment>
-						<RichText
-							tagName="span"
-							className="ab-block-newsletter-label"
-							keepPlaceholderOnFocus
-							formattingControls={ [] }
-							value={ attributes.emailInputLabel }
-							onChange={ ( value ) => this.props.setAttributes( { emailInputLabel: value } ) }
-						/>
-
-						<TextControl
-							name="ab-newsletter-email-address"
-						/>
-
-						<div
-							className={ editClassName ? editClassName : undefined }
-							style={ editStyles }
-						>
-							<CustomButton { ...this.props }>
-								<RichText
-									tagName="span"
-									placeholder={ __( 'Button text...', 'atomic-blocks' ) }
-									keepPlaceholderOnFocus
-									value={ attributes.buttonText }
-									formattingControls={ [] }
-									className={ classnames(
-										'ab-button',
-										attributes.buttonClass,
-										attributes.buttonShape,
-										attributes.buttonSize,
-										getButtonBackgroundClass,
-										getButtonTextClass,
-										{
-											'has-background': attributes.buttonBackgroundColor || attributes.customButtonBackgroundColor,
-											'has-text-color': attributes.buttonTextColor || attributes.customButtonTextColor,
-										}
-									) }
-									style={ {
-										backgroundColor: getButtonBackgroundClass ? undefined : attributes.customButtonBackgroundColor,
-										color: getButtonTextClass ? undefined : attributes.customButtonTextColor,
-									} }
-									onChange={ (value) => this.props.setAttributes( { buttonText: value } ) }
-								/>
-							</CustomButton>
-							{ isSelected && (
-								<form
-									key="form-link"
-									className={ `blocks-button__inline-link ab-button-${attributes.buttonAlignment}`}
-									onSubmit={ event => event.preventDefault() }
-									style={ {
-										textAlign: attributes.buttonAlignment,
-									} }
-								>
-								</form>
-							) }
-						</div>
-					</Fragment>
-				) }
-			</NewsletterContainer>
-		];
-	}
-}
 
 registerBlockType(
 	'atomic-blocks/newsletter',
@@ -132,11 +21,11 @@ registerBlockType(
 		keywords: [
 			__( 'Mailchimp', 'atomic-blocks' ),
 			__( 'Subscribe', 'atomic-blocks' ),
+			__( 'Newsletter', 'atomic-blocks' ),
 		],
 		edit: Edit,
 		save: () => {
 			return null;
 		}
 	},
-
 );
